test(vue-query): vitest type test correctly
